Refactored to implement context menu with rename and delete across all directory items.
