+++ src/pages/editor/FileBarCompoents/RenameFile.jsx
@@ -2,20 +2,27 @@ import { useFileBarDataProvider } from "@/contexts/FileBarDataProvider";
 import React, { useEffect, useRef, useState } from "react";
 
 function RenameFile({ item }) {
+  console.log("render");
+
   const [name, setName] = useState(item.name || "");
   const { updateFile, deleteFile } = useFileBarDataProvider();
   const ref = useRef(null);
 
+  useEffect(() => {
+    if (ref.current) ref.current.focus();
+  }, []);
+
   const handleClickOutSide = (event) => {
     if (ref.current && !ref.current.contains(event.target)) {
       if (name) updateFile(item.id, { name, type: item.toBeType });
+      else if (item.name)
+        updateFile(item.id, { name: item.name, type: item.toBeType });
       else deleteFile(item.id);
     }
   };
 
   // useEffect(() => {
   //   console.log(name);
-
   //   return () => {
   //     console.log(name);
   //     // deleteFile(item.id);
@@ -37,7 +44,6 @@ function RenameFile({ item }) {
         value={name}
         ref={ref}
         onChange={handleChange}
-        autoFocus
         className="bg-slate-700 outline-none w-full px-1 focus:ring-2"
       />
     </div>
